creacion de dto, entidades, controller , etc iniciales de user service

+++ user-service/src/main/java/com/user/service/servicio/UserService.java
@@ -22,7 +22,7 @@ public class UserService {
 		return userRepository.findAll();
 	}
 	
-	public User obtenerUsurario(int id) {
+	public User obtenerUsurario(long id) {
 		return userRepository.findById(id).orElse(null);
 	}
